// src/enums/exception.ts
export enum ExceptionKeys  {
    LOCAL_NETWORK_ERROR = "LOCAL_NETWORK_ERROR",
    NETWORK_ERROR = "NETWORK_ERROR",
    VALIDATION_ERROR = "VALIDATION_ERROR",
    UNAUTHORIZED_ERROR = "UNAUTHORIZED_ERROR",
    FORBIDDEN_ERROR = "FORBIDDEN_ERROR",
    NOT_FOUND_ERROR = "NOT_FOUND_ERROR",
    INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
    SUCCESS_TITLE = "SUCCESS_TITLE",
    SUCCESS_DESC = "SUCCESS_DESC",
}